% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_redun_sim.R
\name{extinction_simulation}
\alias{extinction_simulation}
\title{Results of functional redundancy simulations}
\usage{
extinction_simulation(
  range.express = c(1:100),
  range.nonexpress = c(0:100),
  q = 0.5,
  a.range = seq(0, 0.4, by = 0.01),
  loss.thres = 0.99,
  range.trait = seq(1, 1000, by = 1),
  n.community = 100,
  n.extinct.sim = 10
)
}
\arguments{
\item{range.express}{The number of taxa contributing to a community-aggregated parameter--vector}

\item{range.nonexpress}{The number of taxa not contributing to a community-aggregated parameter--vector}

\item{q}{The diversity order to use when calculate functional redundancy--real number}

\item{a.range}{A parameter modulating how even a lognormal abundance distribution is--vector}

\item{loss.thres}{A parameter specifying the total loss trait in a community necessary to stop a simulation--number between 0 and 1}

\item{range.trait}{Possible range of trait values to be assigned to taxa--a vector}

\item{n.community}{Number of artificial communities to generate--integer greater than 0}

\item{n.extinct.sim}{Number of extinction simulations to perform on every artificial community--integer greater than 0}
}
\value{
RReturns a list of fr (functional redundancy), loss.frac (average fraction of a community required to go extinct prior to trait loss), and express.level (the fraction of a community required to go extinct)
}
\description{
Performs extinction simulations and calculates functional redundancy using the Royalty method.
}
\details{
ANYTHING THAT NEEDS TO GO INTO THE DETAILS SECTION
}
\examples{
# We should include examples, but I'm kind of confused about how examples work
}
