% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_redun_sim.R
\docType{data}
\name{func_redun_sim}
\alias{func_redun_sim}
\title{Results of functional redundancy simulations}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 50 rows and 4 columns.
}
\usage{
func_redun_sim
}
\description{
Can put a paragraph of explanation here
}
\details{
#' @format A data frame with four variables: \code{fr}, \code{method},
  \code{loss.frac}, and \code{express.level}
}
\keyword{datasets}
