% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hester_fr.R
\name{hester_fr}
\alias{hester_fr}
\title{Functional redundancy via Hester's method}
\usage{
hester_fr(trait.contribution)
}
\arguments{
\item{trait.contribution}{#DESCRIBE THIS HERE}
}
\description{
Functional redundancy via Hester's method
}
