% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricotta_fr.R
\name{ricotta_fr}
\alias{ricotta_fr}
\title{Functional redundancy via Ricotta's method}
\usage{
ricotta_fr(p, trait)
}
\arguments{
\item{p}{# WHAT IS THIS}

\item{trait}{# WHAT IS THIS}
}
\description{
Functional redundancy via Ricotta's method
}
