% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species.abund}
\alias{species.abund}
\title{Abundances of OTUs in the TARA dataset}
\format{
A large matrix (75 rows, 2451 columns, ~1.7 MB)
\describe{
  \item{}{}
  \item{}{}
  ...
}

A matrix (2631 rows, 8 columns, ~275 kB)
\describe{
  \item{}{}
  \item{}{}
  ...
}
}
\source{
\url{}

\url{}
}
\usage{
species.abund

species.abund
}
\description{
A dataset containing the abundances of OTUs in the combined TARA dataset

A dataset containing the abundances trait levels for each OTU in the TARA dataset, for 7 genes related to N cycling
}
\keyword{datasets}
