% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royalty_fr.R
\name{royalty_fr}
\alias{royalty_fr}
\title{Functional redundancy via Royalty's method}
\usage{
royalty_fr(trait.contribution, q.range = seq(0, 2, by = 0.1))
}
\arguments{
\item{trait.contribution}{DESCRIBE THIS}

\item{q.range}{DESCRIBE THIS}
}
\value{
EXPLAIN WHAT IT RETURNS, WITH OBJECT TYPE
}
\description{
Calculates functional redundancy via the Royalty method, as described in INSERT REFERENCE HERE.
}
\details{
ANYTHING THAT NEEDS TO GO INTO THE DETAILS SECTION
}
\examples{
# We should include examples, but I'm kind of confused about how examples work
}
\references{
ADD MANUSCRIPT REFERENCE HERE
}
